requirements.txt
pyyaml>=6.0
requests>=2.28.0

# For 'make menuconfig' (same ncurses UI as compressor_2)
kconfiglib
